backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
clickhouse-driver==0.2.5
python-multipart==0.0.5
python-jose[cryptography]==3.3.0
pandas==1.3.3
pydantic==1.8.2
python-dotenv==0.19.0
aiofiles==0.7.0
pytest==6.2.5
httpx==0.19.0 